// api/model/roamingSessionDto.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreditCardDto } from './creditCardDto';
import { RoamingTariffDto } from './roamingTariffDto';
import { RoamingSessionStatus } from './roamingSessionStatus';
import { RoamingPointDto } from './roamingPointDto';
import { ChargeSessionData } from './chargeSessionData';
import { RoamingConnectorDto } from './roamingConnectorDto';


export interface RoamingSessionDto { 
    id?: string | null;
    userId?: string | null;
    chargePointInterestId?: string | null;
    chargePointInterestSocketId?: string | null;
    connectorNo?: number;
    status?: RoamingSessionStatus;
    externalSessionId?: string | null;
    externalCPOSessionId?: string | null;
    transactionId?: string | null;
    no?: number;
    readableId?: string | null;
    tariffId?: string | null;
    startReason?: string | null;
    stopReason?: string | null;
    creditCardId?: string | null;
    readonly plateNo?: string | null;
    chargeSessionData?: ChargeSessionData;
    isArchived?: boolean;
    archiveDate?: string | null;
    archivedBy?: string | null;
    creditCard?: CreditCardDto;
    tariff?: RoamingTariffDto;
    chargePointInterest?: RoamingPointDto;
    chargePointInterestSocket?: RoamingConnectorDto;
    startDate?: string | null;
    endDate?: string | null;
    partyId?: string | null;
    provider?: string | null;
}
export namespace RoamingSessionDto {
}